wait i was using an outdated version somehow

+++ admin/modules/forum/moderation_queue.php
@@ -1,12 +1,10 @@
 <?php
 /**
- * MyBB 1.6
- * Copyright 2010 MyBB Group, All Rights Reserved
+ * MyBB 1.8
+ * Copyright 2014 MyBB Group, All Rights Reserved
  *
- * Website: http://mybb.com
- * License: http://mybb.com/about/license
- *
- * $Id$
+ * Website: http://www.mybb.com
+ * License: http://www.mybb.com/about/license
  */
 
 // Disallow direct access to this file for security reasons
@@ -41,13 +39,14 @@ $plugins->run_hooks("admin_forum_moderation_queue_begin");
 if($mybb->request_method == "post")
 {
 	$plugins->run_hooks("admin_forum_moderation_queue_commit");
-	
+
 	require_once MYBB_ROOT."inc/functions_upload.php";
 	require_once MYBB_ROOT."inc/class_moderation.php";
 	$moderation = new Moderation;
 
 	if(is_array($mybb->input['threads']))
 	{
+		$threads_to_approve = $threads_to_delete = array();
 		// Fetch threads
 		$query = $db->simple_select("threads", "tid", "tid IN (".implode(",", array_map("intval", array_keys($mybb->input['threads'])))."){$flist}");
 		while($thread = $db->fetch_array($query))
@@ -57,16 +56,24 @@ if($mybb->request_method == "post")
 			{
 				$threads_to_approve[] = $thread['tid'];
 			}
-			else if($action == "delete")
+			else if($action == "delete" && $mybb->settings['soft_delete'] != 1)
 			{
 				$moderation->delete_thread($thread['tid']);
 			}
+			else if($action == "delete")
+			{
+				$threads_to_delete[] = $thread['tid'];
+			}
 		}
-		if(is_array($threads_to_approve))
+		if(!empty($threads_to_approve))
 		{
 			$moderation->approve_threads($threads_to_approve);
 		}
-		
+		if(!empty($threads_to_delete))
+		{
+			$moderation->soft_delete_threads($threads_to_delete);
+		}
+
 		$plugins->run_hooks("admin_forum_moderation_queue_threads_commit");
 
 		// Log admin action
@@ -77,6 +84,7 @@ if($mybb->request_method == "post")
 	}
 	else if(is_array($mybb->input['posts']))
 	{
+		$posts_to_approve = $posts_to_delete = array();
 		// Fetch posts
 		$query = $db->simple_select("posts", "pid", "pid IN (".implode(",", array_map("intval", array_keys($mybb->input['posts'])))."){$flist}");
 		while($post = $db->fetch_array($query))
@@ -86,16 +94,24 @@ if($mybb->request_method == "post")
 			{
 				$posts_to_approve[] = $post['pid'];
 			}
-			else if($action == "delete")
+			else if($action == "delete" && $mybb->settings['soft_delete'] != 1)
 			{
 				$moderation->delete_post($post['pid']);
 			}
+			else if($action == "delete")
+			{
+				$posts_to_delete[] = $post['pid'];
+			}
 		}
-		if(is_array($posts_to_approve))
+		if(!empty($posts_to_approve))
 		{
 			$moderation->approve_posts($posts_to_approve);
 		}
-		
+		if(!empty($posts_to_delete))
+		{
+			$moderation->soft_delete_posts($posts_to_delete);
+		}
+
 		$plugins->run_hooks("admin_forum_moderation_queue_posts_commit");
 
 		// Log admin action
@@ -120,7 +136,7 @@ if($mybb->request_method == "post")
 				remove_attachment($attachment['pid'], '', $attachment['aid']);
 			}
 		}
-		
+
 		$plugins->run_hooks("admin_forum_moderation_queue_attachments_commit");
 
 		// Log admin action
@@ -132,16 +148,16 @@ if($mybb->request_method == "post")
 }
 
 $all_options = "<ul class=\"modqueue_mass\">\n";
-$all_options .= "<li><a href=\"#\" class=\"mass_ignore\" onclick=\"$$('input.radio_ignore').each(function(e) { e.checked = true; }); return false;\">{$lang->mark_as_ignored}</a></li>\n";
-$all_options .= "<li><a href=\"#\" class=\"mass_delete\" onclick=\"$$('input.radio_delete').each(function(e) { e.checked = true; }); return false;\">{$lang->mark_as_deleted}</a></li>\n";
-$all_options .= "<li><a href=\"#\" class=\"mass_approve\" onclick=\"$$('input.radio_approve').each(function(e) { e.checked = true; }); return false;\">{$lang->mark_as_approved}</a></li>\n";
+$all_options .= "<li><a href=\"#\" class=\"mass_ignore\">{$lang->mark_as_ignored}</a></li>\n";
+$all_options .= "<li><a href=\"#\" class=\"mass_delete\">{$lang->mark_as_deleted}</a></li>\n";
+$all_options .= "<li><a href=\"#\" class=\"mass_approve\">{$lang->mark_as_approved}</a></li>\n";
 $all_options .= "</ul>\n";
 
 // Threads awaiting moderation
 if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 {
 	$plugins->run_hooks("admin_forum_moderation_queue_threads");
-	
+
 	$forum_cache = $cache->read("forums");
 
 	$query = $db->simple_select("threads", "COUNT(tid) AS unapprovedthreads", "visible=0");
@@ -153,9 +169,9 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 		$per_page = 15;
 		if($mybb->input['page'] > 0)
 		{
-			$current_page = intval($mybb->input['page']);
+			$current_page = $mybb->get_input('page', MyBB::INPUT_INT);
 			$start = ($current_page-1)*$per_page;
-			$pages = $unaproved_threads / $per_page;
+			$pages = $unapproved_threads / $per_page;
 			$pages = ceil($pages);
 			if($current_page > $pages)
 			{
@@ -169,7 +185,7 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 			$current_page = 1;
 		}
 
-		$pagination = draw_admin_pagination($current_page, $per_page, $unaproved_threads, "index.php?module=forum-moderation_queue&amp;page={page}");
+		$pagination = draw_admin_pagination($current_page, $per_page, $unapproved_threads, "index.php?module=forum-moderation_queue&amp;page={page}");
 
 		$page->add_breadcrumb_item($lang->threads_awaiting_moderation);
 		$page->output_header($lang->threads_awaiting_moderation);
@@ -183,7 +199,7 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 		$table->construct_header($lang->posted, array("class" => "align_center", "width" => "20%"));
 
 		$query = $db->query("
-			SELECT t.tid, t.dateline, t.fid, t.subject, p.message AS postmessage, u.username AS username, t.uid
+			SELECT t.tid, t.dateline, t.fid, t.subject, t.username AS threadusername, p.message AS postmessage, u.username AS username, t.uid
 			FROM ".TABLE_PREFIX."threads t
 			LEFT JOIN ".TABLE_PREFIX."posts p ON (p.pid=t.firstpost)
 			LEFT JOIN ".TABLE_PREFIX."users u ON (u.uid=t.uid)
@@ -197,14 +213,29 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 			$thread['threadlink'] = get_thread_link($thread['tid']);
 			$thread['forumlink'] = get_forum_link($thread['fid']);
 			$forum_name = $forum_cache[$thread['fid']]['name'];
-			$threaddate = my_date($mybb->settings['dateformat'], $thread['dateline']);
-			$threadtime = my_date($mybb->settings['timeformat'], $thread['dateline']);
-			$profile_link = build_profile_link($thread['username'], $thread['uid'], "_blank");
+			$threaddate = my_date('relative', $thread['dateline']);
+
+			if(!$thread['uid'])
+			{
+				if($thread['threadusername'] != "")
+				{
+					$profile_link = $thread['threadusername'];
+				}
+				else
+				{
+					$profile_link = htmlspecialchars_uni($lang->guest);
+				}
+			}
+			else
+			{
+				$profile_link = build_profile_link(htmlspecialchars_uni($thread['username']), $thread['uid'], "_blank");
+			}
+
 			$thread['postmessage'] = nl2br(htmlspecialchars_uni($thread['postmessage']));
 
 			$table->construct_cell("<a href=\"../{$thread['threadlink']}\" target=\"_blank\">{$thread['subject']}</a>");
 			$table->construct_cell($profile_link, array("class" => "align_center"));
-			$table->construct_cell("{$threaddate}, {$threadtime}", array("class" => "align_center"));
+			$table->construct_cell($threaddate, array("class" => "align_center"));
 			$table->construct_row();
 
 			$controls = "<div class=\"modqueue_controls\">\n";
@@ -219,13 +250,35 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 			$table->construct_row();
 		}
 
-		$table->output($lang->threads_awaiting_moderation);
+		$table->output($lang->threads_awaiting_moderation, 1, "general tfixed");
 		echo $all_options;
 		echo $pagination;
 
 		$buttons[] = $form->generate_submit_button($lang->perform_action);
 		$form->output_submit_wrapper($buttons);
 		$form->end();
+
+		echo '<script type="text/javascript">
+			$(".mass_ignore").on("click", function () {
+				$("input.radio_ignore").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_delete").on("click", function () {
+				$("input.radio_delete").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_approve").on("click", function () {
+				$("input.radio_approve").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+		</script>';
+
 		$page->output_footer();
 	}
 }
@@ -234,7 +287,7 @@ if($mybb->input['type'] == "threads" || !$mybb->input['type'])
 if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 {
 	$plugins->run_hooks("admin_forum_moderation_queue_posts");
-	
+
 	$forum_cache = $cache->read("forums");
 
 	$query = $db->query("
@@ -251,9 +304,9 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 		$per_page = 15;
 		if($mybb->input['page'] > 0)
 		{
-			$current_page = intval($mybb->input['page']);
+			$current_page = $mybb->get_input('page', MyBB::INPUT_INT);
 			$start = ($current_page-1)*$per_page;
-			$pages = $unaproved_posts / $per_page;
+			$pages = $unapproved_posts / $per_page;
 			$pages = ceil($pages);
 			if($current_page > $pages)
 			{
@@ -267,7 +320,7 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 			$current_page = 1;
 		}
 
-		$pagination = draw_admin_pagination($current_page, $per_page, $unaproved_posts, "index.php?module=forum-moderation_queue&amp;type=posts&amp;page={page}");
+		$pagination = draw_admin_pagination($current_page, $per_page, $unapproved_posts, "index.php?module=forum-moderation_queue&amp;type=posts&amp;page={page}");
 
 
 		$page->add_breadcrumb_item($lang->posts_awaiting_moderation);
@@ -282,7 +335,7 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 		$table->construct_header($lang->posted, array("class" => "align_center", "width" => "20%"));
 
 		$query = $db->query("
-			SELECT p.pid, p.subject, p.message, t.subject AS threadsubject, t.tid, u.username, p.uid, t.fid
+			SELECT p.pid, p.subject, p.message, p.dateline, p.username AS postusername, t.subject AS threadsubject, t.tid, u.username, p.uid, t.fid
 			FROM  ".TABLE_PREFIX."posts p
 			LEFT JOIN ".TABLE_PREFIX."threads t ON (t.tid=p.tid)
 			LEFT JOIN ".TABLE_PREFIX."users u ON (u.uid=p.uid)
@@ -295,7 +348,7 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 			$altbg = alt_trow();
 			$post['threadsubject'] = htmlspecialchars_uni($post['threadsubject']);
 			$post['subject'] = htmlspecialchars_uni($post['subject']);
-			
+
 			if(!$post['subject'])
 			{
 				$post['subject'] = $lang->re." ".$post['threadsubject'];
@@ -305,14 +358,29 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 			$post['threadlink'] = get_thread_link($post['tid']);
 			$post['forumlink'] = get_forum_link($post['fid']);
 			$forum_name = $forum_cache[$post['fid']]['name'];
-			$postdate = my_date($mybb->settings['dateformat'], $post['dateline']);
-			$posttime = my_date($mybb->settings['timeformat'], $post['dateline']);
-			$profile_link = build_profile_link($post['username'], $post['uid'], "_blank");
+			$postdate = my_date('relative', $post['dateline']);
+
+			if(!$post['uid'])
+			{
+				if($post['postusername'] != "")
+				{
+					$profile_link = $post['postusername'];
+				}
+				else
+				{
+					$profile_link = $lang->guest;
+				}
+			}
+			else
+			{
+				$profile_link = build_profile_link(htmlspecialchars_uni($post['username']), $post['uid'], "_blank");
+			}
+
 			$post['message'] = nl2br(htmlspecialchars_uni($post['message']));
 
 			$table->construct_cell("<a href=\"../{$post['postlink']}#pid{$post['pid']}\" target=\"_blank\">{$post['subject']}</a>");
 			$table->construct_cell($profile_link, array("class" => "align_center"));
-			$table->construct_cell("{$postdate}, {$posttime}", array("class" => "align_center"));
+			$table->construct_cell($postdate, array("class" => "align_center"));
 			$table->construct_row();
 
 			$controls = "<div class=\"modqueue_controls\">\n";
@@ -328,13 +396,35 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 			$table->construct_row();
 		}
 
-		$table->output($lang->posts_awaiting_moderation);
+		$table->output($lang->posts_awaiting_moderation, 1, "general tfixed");
 		echo $all_options;
 		echo $pagination;
 
 		$buttons[] = $form->generate_submit_button($lang->perform_action);
 		$form->output_submit_wrapper($buttons);
 		$form->end();
+
+		echo '<script type="text/javascript">
+			$(".mass_ignore").on("click", function () {
+				$("input.radio_ignore").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_delete").on("click", function () {
+				$("input.radio_delete").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_approve").on("click", function () {
+				$("input.radio_approve").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+		</script>';
+
 		$page->output_footer();
 	}
 	else if($mybb->input['type'] == "posts")
@@ -350,7 +440,7 @@ if($mybb->input['type'] == "posts" || $mybb->input['type'] == "")
 if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 {
 	$plugins->run_hooks("admin_forum_moderation_queue_attachments");
-	
+
 	$query = $db->query("
 		SELECT COUNT(aid) AS unapprovedattachments
 		FROM  ".TABLE_PREFIX."attachments a
@@ -366,7 +456,7 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 		$per_page = 15;
 		if($mybb->input['page'] > 0)
 		{
-			$current_page = intval($mybb->input['page']);
+			$current_page = $mybb->get_input('page', MyBB::INPUT_INT);
 			$start = ($current_page-1)*$per_page;
 			$pages = $unapproved_attachments / $per_page;
 			$pages = ceil($pages);
@@ -397,7 +487,7 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 		$table->construct_header($lang->controls, array("class" => "align_center", "colspan" => 3));
 
 		$query = $db->query("
-			SELECT a.*, p.subject AS postsubject, p.dateline, p.uid, u.username, t.tid, t.subject AS threadsubject
+			SELECT a.*, p.subject AS postsubject, p.dateline, p.username AS postusername, p.uid, u.username, t.tid, t.subject AS threadsubject
 			FROM  ".TABLE_PREFIX."attachments a
 			LEFT JOIN ".TABLE_PREFIX."posts p ON (p.pid=a.pid)
 			LEFT JOIN ".TABLE_PREFIX."threads t ON (t.tid=p.tid)
@@ -410,8 +500,7 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 		while($attachment = $db->fetch_array($query))
 		{
 			if(!$attachment['dateuploaded']) $attachment['dateuploaded'] = $attachment['dateline'];
-			$attachdate = my_date($mybb->settings['dateformat'], $attachment['dateuploaded']);
-			$attachtime = my_date($mybb->settings['timeformat'], $attachment['dateuploaded']);
+			$attachdate = my_date('relative', $attachment['dateuploaded']);
 
 			$attachment['postsubject'] = htmlspecialchars_uni($attachment['postsubject']);
 			$attachment['filename'] = htmlspecialchars_uni($attachment['filename']);
@@ -420,11 +509,26 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 
 			$link = get_post_link($attachment['pid'], $attachment['tid']) . "#pid{$attachment['pid']}";
 			$thread_link = get_thread_link($attachment['tid']);
-			$profile_link = build_profile_link($attachment['username'], $attachment['uid'], "_blank");
+
+			if(!$attachment['uid'])
+			{
+				if($attachment['postusername'] != "")
+				{
+					$profile_link = $attachment['postusername'];
+				}
+				else
+				{
+					$profile_link = htmlspecialchars_uni($lang->guest);
+				}
+			}
+			else
+			{
+				$profile_link = build_profile_link(htmlspecialchars_uni($attachment['username']), $attachment['uid'], "_blank");
+			}
 
 			$table->construct_cell("<a href=\"../attachment.php?aid={$attachment['aid']}\" target=\"_blank\">{$attachment['filename']}</a> ({$attachment['filesize']})<br /><small class=\"modqueue_meta\">{$lang->post} <a href=\"{$link}\" target=\"_blank\">{$attachment['postsubject']}</a></small>");
 			$table->construct_cell($profile_link, array("class" => "align_center"));
-			$table->construct_cell("{$attachdate}, {$attachtime}", array("class" => "align_center"));
+			$table->construct_cell($attachdate, array("class" => "align_center"));
 
 			$table->construct_cell($form->generate_radio_button("attachments[{$attachment['aid']}]", "ignore", $lang->ignore, array('class' => 'radio_ignore', 'checked' => true)), array("class" => "align_center"));
 			$table->construct_cell($form->generate_radio_button("attachments[{$attachment['aid']}]", "delete", $lang->delete, array('class' => 'radio_delete', 'checked' => false)), array("class" => "align_center"));
@@ -438,6 +542,28 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 		$buttons[] = $form->generate_submit_button($lang->perform_action);
 		$form->output_submit_wrapper($buttons);
 		$form->end();
+
+		echo '<script type="text/javascript">
+			$(".mass_ignore").on("click", function () {
+				$("input.radio_ignore").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_delete").on("click", function () {
+				$("input.radio_delete").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+			$(".mass_approve").on("click", function () {
+				$("input.radio_approve").each(function(e) {
+					$(this).prop("checked", true);
+				});
+				return false;
+			});
+		</script>';
+
 		$page->output_footer();
 	}
 	else if($mybb->input['type'] == "attachments")
@@ -453,5 +579,3 @@ if($mybb->input['type'] == "attachments" || $mybb->input['type'] == "")
 $page->output_header($lang->moderation_queue);
 echo "<p class=\"notice\">{$lang->error_no_threads}</p>";
 $page->output_footer();
-
-?>
